concept for external commands
